==> ames-housing-prices/04_Production_Model_and_Insights.ipynb <==
### Objective:
   1. Build the final model for predicting prices
   

import pandas as pd
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
import seaborn as sns
import sys
from sklearn.linear_model import LinearRegression, LassoCV, RidgeCV, Ridge
from sklearn.neighbors import KNeighborsClassifier
from sklearn.preprocessing import PolynomialFeatures, StandardScaler
from sklearn.model_selection import train_test_split, cross_val_score, cross_val_predict
from sklearn.metrics import r2_score
# insert at 1, 0 is the script path (or '' in REPL)
sys.path.append('/Users/ganeshsivam/Mods')
import corr

pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)

sns.set_style('darkgrid')
%config InlineBackend.figure_format = 'retina'
%matplotlib inline

### Run our cleaning, preprocessing, FE  & Model tuning notebook to import the functions & models needed

%run 01_EDA_and_Cleaning.ipynb
plt.ioff()

%run 02_Preprocessing_and_Feature_Engineering.ipynb

%run 03_Model_Benchmarks_&_Tuning_&_Subimissions.ipynb

### Create our final model based on the hyperparameters from the previous notebook. We are going with the 5th model from the last notebook as this fared well on both Public & Private leaderboard test data. Other models were not so consistent, getting very good scores on one leaderboard but not the other

final_mod


rfe_columns = rfe_cols(150)

Function to take test data as input from a csv file and return predictions

def predict_prices(csv):
    data = pd.read_csv(csv)
    data = data_cleaning(data)
    data = eda_proc_final(data)
    for c in rfe_columns:
        if c not in data.columns:
            data[c] = 0
    data = data[rfe_columns]
    data = final_scaler.transform(data)
    output = final_mod.predict(data)
    return output

final_mod.coef_

d = pd.DataFrame(final_mod.coef_,columns=["Coef"],index=rfe_columns)

d.sort_values("Coef")

d.reset_index(inplace=True)

d.loc[d['index'].str.contains("Kitchen Qua")]

d.loc[(d['index'].str.contains("s2"))|(d['index'].str.contains("s3"))]
